Extract supervisor args/options helpers into SDK-free args.mjs
