+++ lib/agent-supervisor/supervisor.mjs
@@ -27,6 +27,7 @@ import net from 'node:net'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { query } from '@anthropic-ai/claude-agent-sdk'
+import { parseArgsForTest, buildQueryOptions } from './args.mjs'
 import { buildMessengerMcpServer } from './sdk-mcp-tools.mjs'
 import { normalizeSdkMessage, shim } from './stream-shim.mjs'
 
@@ -39,65 +40,15 @@ process.stdout.on('error', (err) => {
   throw err
 })
 
-// --- Arg parsing (minimal, no deps) ---
-
-export function parseArgsForTest(argv) {
-  const args = {
-    role: 'agent',
-    idleTimeoutMs: 10 * 60 * 1000,
-    eventLogDir: '/logs/events',
-    ignoreInboxBefore: 0,
-  }
-  for (let i = 2; i < argv.length; i++) {
-    const arg = argv[i]
-    if (arg === '--role') args.role = argv[++i]
-    else if (arg === '--instance') args.instance = argv[++i]
-    else if (arg === '--prompt-file') args.promptFile = argv[++i]
-    else if (arg === '--resume-session') args.resumeSession = argv[++i]
-    else if (arg === '--model') args.model = argv[++i]
-    else if (arg === '--cwd') args.cwd = argv[++i]
-    else if (arg === '--system-prompt-file') args.systemPromptFile = argv[++i]
-    else if (arg === '--append-system-prompt') args.appendSystemPrompt = argv[++i]
-    else if (arg === '--mcp-config') args.mcpConfig = argv[++i]
-    else if (arg === '--idle-timeout-ms') args.idleTimeoutMs = Number.parseInt(argv[++i], 10)
-    else if (arg === '--no-effort-max') args.noEffortMax = true
-    else if (arg === '--event-log-dir') args.eventLogDir = argv[++i]
-    else if (arg === '--ignore-inbox-before') args.ignoreInboxBefore = Number.parseInt(argv[++i], 10)
-    else if (arg === '--help' || arg === '-h') {
-      throw new Error('__help__')
-    } else {
-      throw new Error(`unknown argument ${arg}`)
-    }
-  }
-  if (args.resumeSession && args.promptFile) {
-    throw new Error('--resume-session and --prompt-file are mutually exclusive')
-  }
-  return args
-}
-
-/**
- * Build the options object passed to the Claude Agent SDK's query().
- * Pure function — no side effects, no I/O — so it's safely unit-testable.
- * When resumeSession is set, attaches `resume` and the SDK replays the
- * transcript instead of starting from a fresh prompt.
- */
-export function buildQueryOptions({ model, cwd, mcpServers, allowedTools, appendSystemPrompt, resumeSession }) {
-  const options = {
-    model,
-    cwd,
-    permissionMode: 'bypassPermissions',
-    mcpServers,
-    allowedTools,
-    includePartialMessages: false,
-  }
-  if (appendSystemPrompt) {
-    options.appendSystemPrompt = appendSystemPrompt
-  }
-  if (resumeSession) {
-    options.resume = resumeSession
-  }
-  return options
-}
+// --- Arg parsing ---
+//
+// The pure parser (parseArgsForTest) and queryOptions builder
+// (buildQueryOptions) live in ./args.mjs so that `node --test` can
+// exercise them without needing node_modules/ — importing this file
+// would eagerly load the Claude Agent SDK.
+//
+// Re-exported here for backward-compatibility with existing callers.
+export { parseArgsForTest, buildQueryOptions }
 
 function parseArgs(argv) {
   try {
