+++ lib/agent-supervisor/supervisor.test.mjs
@@ -1,6 +1,9 @@
 import { test } from 'node:test'
 import assert from 'node:assert/strict'
-import { parseArgsForTest, buildQueryOptions } from './supervisor.mjs'
+// Import from args.mjs (the SDK-free module) so these tests can run
+// without node_modules/ installed. supervisor.mjs re-exports the same
+// names for runtime callers.
+import { parseArgsForTest, buildQueryOptions } from './args.mjs'
 
 test('parses --resume-session', () => {
   const args = parseArgsForTest(['node', 'supervisor.mjs',
